CryptStage now uses a CryptOpHandler.
The class has been reworked as described in the previous comment.

+++ src/main/scala/com/github/sync/crypt/CryptOpHandler.scala
@@ -51,5 +51,5 @@ trait CryptOpHandler {
     * @param orgSize the original file size
     * @return the size after processing by this handler
     */
-  def processedSize(orgSize: Int): Int
+  def processedSize(orgSize: Long): Long
 }
